Fixed selection logic
