+++ desktop/src/main/kotlin/page/SelectPage.kt
@@ -1,7 +1,7 @@
 package page
 
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
+import androidx.compose.desktop.ui.tooling.preview.Preview
+import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material.Button
 import androidx.compose.material.Text
@@ -11,29 +11,42 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.awt.ComposeWindow
 import kotlinx.coroutines.launch
-import selectDirectory
 import java.io.File
+import javax.swing.JFileChooser
+import javax.swing.filechooser.FileNameExtensionFilter
+
+private fun selectDirectory(window: ComposeWindow, title: String): File? {
+	val fileChooser = JFileChooser().apply {
+		dialogTitle = title
+		fileSelectionMode = JFileChooser.DIRECTORIES_ONLY
+		fileFilter = FileNameExtensionFilter("Directory", "directory")
+	}
+
+	val result = fileChooser.showOpenDialog(window)
+	if (result == JFileChooser.APPROVE_OPTION) {
+		return fileChooser.selectedFile
+	}
+	return null
+}
 
 @Composable
-fun SelectPage(window: ComposeWindow, onDirectoriesSelected: (Set<File>) -> Unit) {
-    val coroutineScope = rememberCoroutineScope()
+@Preview
+fun SelectPage(window: ComposeWindow, onDirectorySelected: (File?) -> Unit) {
+	val coroutineScope = rememberCoroutineScope()
 
-    Column(
-        modifier = Modifier.fillMaxSize(),
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.Center
-    ) {
-        Button(onClick = {
-            coroutineScope.launch {
-                val chosenDirectories = selectDirectory(window, "Select Directory with Groups")
-                if (!chosenDirectories.isNullOrEmpty()) {
-                    onDirectoriesSelected(chosenDirectories)
-                } else {
-                    // Show dialog if no subdirectories are found
-                }
-            }
-        }) {
-            Text("Select Directory")
-        }
-    }
+	Box(
+		modifier = Modifier.fillMaxSize(),
+	) {
+		Button(
+			modifier = Modifier.align(Alignment.Center),
+			onClick = {
+				coroutineScope.launch {
+					val dir = selectDirectory(window, "Select Directory with Groups")
+					onDirectorySelected(dir)
+				}
+			}
+		) {
+			Text("Select Directory")
+		}
+	}
 }
